handlers/edits.py
from telegram import Update
from telegram.ext import ContextTypes, CallbackQueryHandler

async def menu_approve(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Раздел 'Одобрение и правки'"""
    query = update.callback_query
    await query.answer()
    await query.edit_message_text("✍ Раздел 'Одобрение и правки' (заглушка)")

def get_handler():
    return CallbackQueryHandler(menu_approve, pattern="^menu_approve$")