requirements.txt
# requirements.txt
python-telegram-bot==21.6
SQLAlchemy==2.0.36
python-dotenv==1.0.1